Make SimpleMDE make events
As SimpleMDE is an override of an existing TextArea, we use it for waiting for that event, but we discard everything from it (see line 31). After that we immediately get the content from SimpleMDE using an FFI function and return it.

The initializeSimpleMDE function sets up everything so our widget functions properly.