(59) feature: file explorer for "Open Project" dialog
This pull-request adds a file explorer to the "Open Project" dialog
which allows the users to navigate through their filesystem just like a
file manager, instead of being required to enter the absolute path.
The explorer's path is synced with the input's path, so they both work
at the same time.